creating per-model-type loaders
