+++ src/loaders/api/index.ts
@@ -1,0 +1,17 @@
+import eventsLoader from "./events";
+import instancesLoader from "./instances";
+import itemsLoader from "./items";
+import organizationsLoader from "./organizations";
+import peopleLoader from "./people";
+import placesLoader from "./places";
+import worksLoader from "./works";
+
+export {
+  eventsLoader,
+  instancesLoader,
+  itemsLoader,
+  organizationsLoader,
+  peopleLoader,
+  placesLoader,
+  worksLoader,
+};
