+++ src/loaders/api/people.ts
@@ -1,0 +1,14 @@
+import { fetchModelData } from "./helpers";
+import { AstroIntegrationLogger } from "astro";
+
+const peopleLoader = async (
+  options: {
+    projectId: number | number[] | string | string[];
+    getRelations?: boolean;
+  },
+  logger?: AstroIntegrationLogger
+) => {
+  return fetchModelData({ ...options, model: "people" }, logger);
+};
+
+export default peopleLoader;
